requirements.txt
aio_pika==6.6.1
aiodocker==0.19.1
aiounittest==1.4.0
motor==2.2.0
pymongo==3.11.0
PyYAML==5.3.1
